stream-service/requirements.txt
fastapi==0.75.0
uvicorn[standard]
websocket-client